services_req-reciever_1746499021.410000/requirements.txt
functions-framework==3.*
pyarrow
flask
google-cloud-bigquery
pandas
pandas-gbq